Refactor HTMLElement: Removed HTMLRenderer
