+++ ts/masters/highcharts.src.ts
@@ -18,7 +18,6 @@ import RendererUtilities from '../Core/Renderer/RendererUtilities.js';
 import SVGElement from '../Core/Renderer/SVG/SVGElement.js';
 import SVGRenderer from '../Core/Renderer/SVG/SVGRenderer.js';
 import HTMLElement from '../Core/Renderer/HTML/HTMLElement.js';
-import HTMLRenderer from '../Core/Renderer/HTML/HTMLRenderer.js';
 import Axis from '../Core/Axis/Axis.js';
 import DateTimeAxis from '../Core/Axis/DateTimeAxis.js';
 import LogarithmicAxis from '../Core/Axis/LogarithmicAxis.js';
@@ -80,8 +79,7 @@ G.Tooltip = Tooltip;
 G.Color = Color;
 G.color = Color.parse;
 // Compositions
-HTMLRenderer.compose(SVGRenderer);
-// HTMLElement.compose(SVGElement);
+HTMLElement.compose(SVGRenderer);
 Pointer.compose(Chart);
 Legend.compose(Chart);
 // DefaultOptions
